Implement RFE #538: Option for stack pointer startup value for stm8.
git-svn-id: http://svn.code.sf.net/p/sdcc/code/trunk@10065 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ sdcc/src/SDCCargs.h
@@ -62,6 +62,6 @@ typedef struct
   } OPTION;
 
 char *getStringArg(const char *szStart, char **argv, int *pi, int argc);
-int getIntArg(const char *szStart, char **argv, int *pi, int argc);
+long getIntArg(const char *szStart, char **argv, int *pi, int argc);
 
 #endif
